% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Hamming distance between two hex-encoded perceptual hashes}
\usage{
hamming_distance(hash1, hash2)
}
\arguments{
\item{hash1}{A character scalar or vector of hex-encoded hashes.}

\item{hash2}{A character scalar or vector of hex-encoded hashes, the
same length as \code{hash1} (or length 1, recycled).}
}
\value{
An integer vector of Hamming distances (number of differing bits).
}
\description{
Computes the number of differing bits between two perceptual hashes
encoded as hexadecimal strings (as produced by \code{\link[=compute_hashes]{compute_hashes()}}).
}
\examples{
hamming_distance("ff00", "ff01")
hamming_distance(c("ffff", "0000"), "0000")

}
